+++ CHANGES.md
@@ -12,6 +12,8 @@ Features:
 * Save and load annotations by name
   * Double-click annotation to rename
 
+* Display interaction handles allowing to rotate or translate current annotation. See [#141](https://github.com/BICCN/cell-locator/issues/141) and [#118](https://github.com/BICCN/cell-locator/issues/118) 
+
 Fixes:
 
 * Drawing a second polygon results in an empty json file.  See [#108](https://github.com/BICCN/cell-locator/issues/108)